funtion EventItem to EventItem.js
