+++ client/src/components/Home.js
@@ -85,332 +85,6 @@ function Details(props) {
     </div>
   );
 }
-
-function EventItem(props) {
-
-  const [editedEvent, setEditedEvent] = useState({});
-  const [startTimeError, setStartTimeError] = useState(false);
-  const [endTimeError, setEndTimeError] = useState(false);
-  const [joinDeadlineError, setJoinDeadlineError] = useState(false);
-
-  const [attending, setAttending] = useState(props.attending);
-  const [attendees, setAttendees] = useState(props.attendees);
-
-  // These states store the original event data
-  const [title, setTitle] = useState(props.title);
-  const [startDate, setStartDate] = useState(props.startDate);
-  const [endDate, setEndDate] = useState(props.endDate);
-  const [joinDeadline, setJoinDeadlineDate] = useState(props.joinDeadline);
-  const [location, setLocation] = useState(props.location);
-  const [description, setDescription] = useState(props.description);
-  const [price, setPrice] = useState(props.price);
-  const [tickets, setTickets] = useState(props.tickets);
-
-  const [checkedTicket, setCheckedTicket] = useState(props.tickets !== 0 && props.tickets !== null && props.tickets !== undefined);
-  const [checkedDeadline, setCheckedDeadline] = useState(!(props.joinDeadline !== null && props.joinDeadline !== undefined));
-
-  // These states store the data that is edited
-  const [edit, setEdit] = useState(false);
-  const [openAttend, setOpenAttend] = useState(false);
-  const [openCancelAttendance, setOpenCancelAttendance] = useState(false);
-  
-
-const showToastMessage = (message) =>
-{
-    toast.error(message, {
-        position: "top-center",
-        autoClose: 6000,
-        hideProgressBar: true,
-        closeOnClick: false,
-        pauseOnHover: false,
-        draggable: false,
-        progress: undefined,
-        theme: "dark"
-        });
-}
-const showToastMessageSuccesfull = (message) =>
-{
-    toast.success(message, {
-        position: "top-center",
-        autoClose: 6000,
-        hideProgressBar: true,
-        closeOnClick: false,
-        pauseOnHover: false,
-        draggable: false,
-        progress: undefined,
-        theme: "light"
-        });
-}
-
-  // POST edited event
-  const saveEditedEventOnClick = (e) => {
-    e.preventDefault()
-    if(!startTimeError && editedEvent.startDate) {
-      const startDateAux = new Date(editedEvent.startDate);
-      if(!endTimeError) {
-        if(!checkedDeadline) {
-          editedEvent.joinDeadline = editedEvent.startDate;
-        }
-        const joinDeadlineAux = new Date(editedEvent.joinDeadline);
-        if(!joinDeadlineError && editedEvent.joinDeadline && joinDeadlineAux <= startDateAux) {
-          const endDateAux = new Date(editedEvent.endDate);
-          if(!editedEvent.endDate || startDateAux < endDateAux) {
-            editedEvent.creator = props.currentUser.firstname + " " + props.currentUser.lastname;
-            editedEvent.creatorId = props.currentUser.id;
-            editedEvent.attendees = 1;
-            editedEvent.id = props.eventID;
-            
-            setEditedEvent(editedEvent);
-            fetch("/api/editEvent", { // TODO change url to event edit url
-              method: "POST",
-              headers: {
-                  "Content-type": "application/json",
-                  "Authorization": "Bearer " + props.currentUser.token
-              },
-              body: JSON.stringify(editedEvent),
-              mode: "cors"
-          })
-              .then(response =>{
-                  if(response.status === 409){
-                    console.log("tickets should be more than the tickets already sold");
-                    showToastMessage("tickets should be more than the tickets already sold");
-                  }
-                  return response.json()
-              } )
-              .then(data => {
-                if (data) { 
-                  showToastMessageSuccesfull(`${data.event.ticketsSold} users will be notified of the changes`);
-                }
-              })
-
-          // Empty the input fields
-          setEditedEvent({});
-          // Close the Modal
-          setEdit(false);
-          if(checkedTicket) {
-            resetTickets();
-          }
-          if(checkedDeadline) {
-            handleDeadlineSwitch();
-          }
-          } else {
-            console.log("Starting time needs to be before ending time.");
-            showToastMessage("Starting time needs to be before ending time.");
-          }
-        } else {
-          console.log("Event joining deadline needs to be before or the same as the starting time of the event.");
-            showToastMessage("Event joining deadline needs to be before or the same as the starting time of the event.");
-        }
-      } else {
-        console.log("Ending time is not valid.");
-        showToastMessage("Ending time is not valid.");
-      }
-    } else {
-      console.log("Starting time is not valid.");
-      showToastMessage("Starting time is not valid.");
-    }
-  };
-
-
-//Resets the amount of tickets and clears the text box.
-const resetTickets = () => {
-  setTickets("");
-  setCheckedTicket(!checkedTicket);
-  editedEvent.tickets = 0;
-  setEditedEvent(editedEvent);
-}
-
-const handleDeadlineSwitch = () => {
-  setCheckedDeadline(!checkedDeadline);
-  if(!checkedDeadline) {
-    editedEvent.joinDeadline = "";
-  } else {
-    editedEvent.joinDeadline = editedEvent.startDate;
-  }
-  
-  setEditedEvent(editedEvent);
-}
-
-//Updates the values for the text fields in event creation
-const whenChanging = (event) => {
-  setEditedEvent({...editedEvent, [event.target.id]: event.target.value})
-}
-
-//Updates the values for the start time
-const handleStartTimeChange = (value) => {
-  setStartTimeError(false); // Sets error to false when changes are made
-  setEditedEvent({...editedEvent, ["startDate"]: value});
-};
-
-//Updates the values for the end time
-const handleEndTimeChange = (value) => {
-  setEndTimeError(false); // Sets error to false when changes are made
-  setEditedEvent({...editedEvent, ["endDate"]: value});
-};
-
-//Updates the values for the join deadline
-const handleJoinDeadlineChange = (value) => {
-  setJoinDeadlineError(false); // Sets error to false when changes are made
-  setEditedEvent({...editedEvent, ["joinDeadline"]: value});
-};
-
-const cancelEditOnClick = () => {
-  setEditedEvent({});
-  if(checkedTicket) {
-    resetTickets();
-  }
-  if(checkedDeadline) {
-    handleDeadlineSwitch();
-  }
-  setEdit(false);
-};
-
-// Triggered if there is an error in the date formatting
-const handleStartTimeError = (error) => {
-  console.log("Starting time error: " + error);
-  setStartTimeError(true);
-}
-
-const handleEndTimeError = (error) => {
-  console.log("Ending time error: " + error);
-  setEndTimeError(true);
-}
-
-const handleJoinDeadlineError = (error) => {
-  console.log("Join deadline error: " + error);
-  setJoinDeadlineError(true);
-} 
-  // Show edit
-  const editOnClick = () => {
-    // Set editedEvent to contain the original event data
-    setEditedEvent({
-      title: title,
-      startDate: startDate,
-      endDate: endDate,
-      joinDeadline: joinDeadline,
-      location: location,
-      description: description,
-      price: props.price,
-      tickets : tickets
-    });
-    setEdit(true);
-  };
-
-  const handleEventAttendance = () => {
-    setOpenAttend(false);
-    setAttending(true);
-    setAttendees(attendees + 1);
-  };
-
-  const handleCancelEventAttendance = () => {
-    setOpenCancelAttendance(false);
-    setAttending(false);
-    setAttendees(attendees - 1);
-  };
-
-
-  if(props.loggedIn) {
-    return (
-      <div className="HomeEventItem">
-        <div className="HomeEventTop">
-          <div className="HomeEventPhotoBackground">
-            <img
-              src="https://blogs.lut.fi/newcomers/wp-content/uploads/sites/15/2020/02/talvi-ilma-1-1.jpg"
-              alt="Event"
-            />
-          </div>
-  
-          <div className="HomeEventHeaderArea">
-            <div className="HomeEventTitleAndLocationArea">
-              <h2>{title}</h2>
-              <h3>Created by: {props.creator}</h3>
-              <h3>{startDate}</h3>
-              <h3>{location}</h3>
-            </div>
-  
-            <p>
-              <FaUserGroup style={{ margin: "0 5px 0 0" }} /> {attendees}
-            </p>
-          </div>
-        </div>
-  
-        <div className="HomeEventBottom">
-          <div className="HorizontalSeparator" />
-          <div className="HomeEventDescriptionArea">
-            <p>{description}</p>
-          </div>
-  
-          <div className='HomeEventAttendanceButtonsArea'>
-            {props.admin && 
-              (
-                <Button className='EditEventButton' variant='contained' onClick={editOnClick} >Edit</Button>
-              )
-            }
-            <div>
-              {attending ? 
-                (
-                  <Button variant='outlined' color='primary' onClick={() => {setOpenCancelAttendance(true)}} >Cancel Attendance</Button>
-                )
-                :
-                (
-                  <Button variant='contained' color='primary' onClick={() => {setOpenAttend(true)}} >Attend the Event</Button>
-                )
-              }
-            </div>
-          </div>
-        </div>
-  
-        <EditEventModal
-          edit={edit} // visible
-          checkedTicket={checkedTicket}
-          checkedDeadline={checkedDeadline}
-          tickets={tickets}
-          setCheckedDeadline={setCheckedDeadline}
-          setTickets={setTickets}
-          setCheckedTicket={setCheckedTicket}
-          whenChanging={whenChanging}
-          handleStartTimeChange={handleStartTimeChange}
-          handleEndTimeChange={handleEndTimeChange}
-          handleJoinDeadlineChange={handleJoinDeadlineChange}
-          cancelEditOnClick={cancelEditOnClick}
-          saveEditedEventOnClick={saveEditedEventOnClick}
-          handleStartTimeError={handleStartTimeError}
-          handleEndTimeError={handleEndTimeError}
-          handleJoinDeadlineError={handleJoinDeadlineError}
-          resetTickets={resetTickets}
-          handleDeadlineSwitch={handleDeadlineSwitch}
-          
-          //old values
-          title={title}
-          description={description}
-          location={location}
-          startDate={startDate}
-          endDate={endDate}
-          joinDeadline={joinDeadline}
-          price={price}
-        />
-  
-        <ConfirmAttendanceModal
-          openAttend={openAttend}
-          setOpenAttend={setOpenAttend}
-          handleEventAttendance={handleEventAttendance}
-        />
-  
-        <CancelAttendanceModal
-          openCancelAttendance={openCancelAttendance}
-          setOpenCancelAttendance={setOpenCancelAttendance}
-          handleCancelEventAttendance={handleCancelEventAttendance}
-        />
-      </div>
-    );
-  } else {
-    return (
-      <Typography sx={{ mt: 20 }} variant='h4' align="center">
-         Please login to view events.
-      </Typography>
-    );
-  }
-  }
   
 
 function Home(props) {
@@ -640,7 +314,6 @@ function Home(props) {
             id={event._id}
             admin={admin}
             eventID={event._id}
-            currentUser={props.currentUser}
             loggedIn={props.currentUser.loggedIn}
             title={event.title}
             creator={event.creator}
@@ -654,8 +327,6 @@ function Home(props) {
             description={event.description}
             tickets={event.tickets}
             ticketsSold={event.ticketsSold}
-            startDate={event.startDate}
-            endDate={event.endDate}
             price={event.price}
             key={index}
             showToastMessage={showToastMessage}
